[feat] : Shot and Cheer @Composable
